+++ pkg/accounts/adaptor/repository/dummy.ts
@@ -3,7 +3,7 @@ import { Option, Result } from '@mikuroxina/mini-fn';
 import { type ID } from '../../../id/type.js';
 import { type Account, type AccountID } from '../../model/account.js';
 import { type AccountFollow } from '../../model/follow.js';
-import type { InactiveAccount } from '../../model/inactiveAccount.js';
+import type { InactiveAccount } from '../../model/inactive.js';
 import type {
   AccountFollowRepository,
   AccountRepository,
