+++ pkg/accounts/service/resendToken.ts
@@ -2,8 +2,8 @@ import { Option, Result } from '@mikuroxina/mini-fn';
 
 import type { AccountName } from '../model/account.js';
 import { type AccountRepository } from '../model/repository.js';
-import type { TokenVerifyService } from './accountVerifyToken.js';
 import { type SendNotificationService } from './sendNotification.js';
+import type { TokenVerifyService } from './verifyToken.js';
 
 export class ResendVerifyTokenService {
   private readonly accountRepository: AccountRepository;
