+++ pkg/accounts/service/verifyToken.ts
